requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.4.1; python_version >= '3.6'

certifi==2021.5.30

cffi==1.14.6

charset-normalizer==2.0.3; python_version >= '3'

coreapi==2.3.3

coreschema==0.0.4

cryptography==3.4.7; python_version >= '3.6'

defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

django-filter==2.4.0

django-templated-mail==1.1.1

django==3.2.5

djangorestframework-simplejwt==4.7.2; python_version >= '3.7'

djangorestframework==3.12.4

djoser==2.1.0

drf-extra-fields==3.1.1

gunicorn==20.1.0

idna==3.2; python_version >= '3'

itypes==1.2.0

jinja2==3.0.1; python_version >= '3.6'

markdown==3.3.4

markupsafe==2.0.1; python_version >= '3.6'

oauthlib==3.1.1; python_version >= '3.6'

pillow==8.3.1

psycopg2-binary==2.9.1

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pyjwt==2.1.0; python_version >= '3.6'

python-dotenv==0.18.0

python3-openid==3.2.0

pytz==2021.1

reportlab==3.5.68

requests-oauthlib==1.3.0

requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

social-auth-app-django==4.0.0

social-auth-core==4.1.0; python_version >= '3.6'

sqlparse==0.4.1; python_version >= '3.5'

uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_full_version < '4.0.0'

